requirements.txt
pandas
pyarrow
boto3
pymongo
python-dotenv
requests
